see 08/17 log

+++ utilcode/src/main/java/com/blankj/utilcode/utils/ConstUtils.java
@@ -18,19 +18,19 @@ public class ConstUtils {
     /**
      * Byte与Byte的倍数
      */
-    public static final long BYTE = 1;
+    public static final int BYTE = 1;
     /**
      * KB与Byte的倍数
      */
-    public static final long KB = 1024;
+    public static final int KB = 1024;
     /**
      * MB与Byte的倍数
      */
-    public static final long MB = 1048576;
+    public static final int MB = 1048576;
     /**
      * GB与Byte的倍数
      */
-    public static final long GB = 1073741824;
+    public static final int GB = 1073741824;
 
     /******************** 时间相关常量 ********************/
     /**
